pyproject.toml
[project]
name = "palate"
version = "0.1.0"
description = "Taiwan restaurant concierge — LLM agent with Google Places tools"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "openai>=2.32.0",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-mock>=3.15.1",
    "respx>=0.23.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live: hits real external APIs; requires env vars; skipped by default",
]
# Skip live tests unless explicitly selected with -m live
addopts = "-m 'not live'"
